Reduce CI load on failing PR
This changes the CI pipeline to run JDK8 first,
and only run the rest of the jobs if it passes.

This will greatly reduce the load for failing PR's,
at the expense of increasing the total time when CI
is not busy and the PR is passing.

It also removes the native image build, and just uses
the output of the standard JDK8 Linux build.