Remove deprecated smallrye-open-api method overrides
The methods `getMethodParameterJaxRsInfo` and `parameterIn` in
`RESTEasyExtension` were meant to add support for RESTEasy annotations
during the OpenAPI scan process. As support for those annotations is now
present in smallrye-open-api and the methods have been deprecated, the
overrides are no longer necessary. This change allows smallrye-open-api
to prune the methods from its API.